+++ dbhose_airflow/move/postgres/delete.sql
@@ -1,23 +1,10 @@
-with 
-    filter_columns as (
-    select
-        array_length(string_to_array('{filter_by}', ', '), 1) as col_count
-)
-  , placeholders as (
-    select
-        string_agg('%L', ', ') as values_placeholders
-    from generate_series(1, (select col_count from filter_columns))
-)
-  , values_list as (
-    select 
-        string_agg(
-            format('(' || (select values_placeholders from placeholders) || ')', {filter_by}), 
-            ', '
-        ) as in_values
-    from {table_temp}
+with filter_columns as (
+    select array_length(string_to_array('{filter_by}', ', '), 1) as col_count
 )
 select
-    true as is_avaliable
-  , 'delete from {table_dest} where ({filter_by}) in (' ||
-    in_values || ')' as move_query
-from values_list
+    true as is_avaliable,
+    'DELETE FROM {table_dest} WHERE EXISTS(' || E'\nSELECT 1 FROM {table_temp}\nWHERE ' || 
+    (select string_agg('{table_dest}.' || trim(col) || ' = {table_temp}.' || trim(col), E'\nAND ') 
+     from unnest(string_to_array('{filter_by}', ',')) as col) || E'\n);\n' ||
+    E'INSERT INTO {table_dest}\nSELECT * FROM {table_temp};' as move_query
+from filter_columns