+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Version History
 
+## 0.0.2.0
+
+* Update depends native-dumper==0.3.3.0
+* Update depends pgpack-dumper==0.3.3.0
+* Update README.md
+* Add create partition into postgres and greenplum ddl queryes
+* Improve delete.sql for greenplum and postgres
+
 ## 0.0.1.0
 
 * Update depends native-dumper==0.3.2.3